Updated tests

+++ auth-service/src/test/java/com/petclinic/auth/User/AuthServiceUserServiceTests.java
@@ -10,7 +10,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.dao.DuplicateKeyException;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.test.context.ActiveProfiles;
 
@@ -48,6 +47,10 @@ public class AuthServiceUserServiceTests {
 
     @BeforeEach
     void setup() {
+        when(mailService.sendMail(any())).then(args -> {
+            final Mail mail = args.getArgument(0, Mail.class);
+            return "Email sent to " + mail.getMessage();
+        });
         userRepo.deleteAllInBatch();
     }
 
@@ -60,6 +63,7 @@ public class AuthServiceUserServiceTests {
         assertEquals(createdUser.getUsername(), userIDLessDTO.getUsername());
         assertNotEquals(createdUser.getPassword(), userIDLessDTO.getPassword());
         assertEquals(createdUser.getEmail(), userIDLessDTO.getEmail());
+        assertFalse(createdUser.isVerified());
     }
 
     @Test
